+++ packages/gallery/src/versionLogger.js
@@ -1,4 +1,4 @@
 try {
-  console.debug('Pro Gallery Version: 3.1.5');
+  console.debug('Pro Gallery Version: 3.1.7');
   // eslint-disable-next-line no-empty
 } catch (e) {}