+++ packages/lib/src/settings/options/titlePlacement.js
@@ -7,9 +7,8 @@ export default {
   isRelevantDescription:
     'Set a Vertical gallery ("Scroll Direction" as "Vertical"), set "Layout Orientation" to "Columns" and set "Max Group Size" to "1".',
   isRelevant: (styleParams) =>
-    styleParams.isVertical &&
-    styleParams.groupSize === 1 &&
-    !styleParams.oneRow,
+    (styleParams.isVertical || styleParams.oneRow) &&
+    styleParams.groupSize === 1,
   type: INPUT_TYPES.MULTISELECT,
   default: GALLERY_CONSTS.placements.SHOW_ON_HOVER,
   options: createOptions('placements'),