+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## [v3.1.7](https://pro-gallery-3-1-7.surge.sh) (07/04/2021)
+ 
+#### LIB
+ -  blueprints: added handling text item changes in itemsChanged.
+
+---
+## [v3.1.6](https://pro-gallery-3-1-6.surge.sh) (06/04/2021)
+ 
+#### GALLERY
+ -  improve (slideshowView): make the arrowSize container responsive
+
+---
 ## [v3.1.5](https://pro-gallery-3-1-5.surge.sh) (05/04/2021)
  
 #### GALLERY